improve first paint time on index page
Waiting to fetch the HTML for the tweets from the database is
unacceptable. For now, I'm storing the first five so they can be
rendered immediately and the response can be sent back to the client.

I think I need to reevaluate whether or not I want to fetch tweet
HTML from the database at all. The latency is noticeable and painful.