don't fetch translation count for each tweet
This was slowing down the time to first paint big time. The app
would request the tweets, then for each one wait to hear back about
how many translations are available for that tweet.

This is a temporary solution. The next step is to store translations
within the tweet documents in the database so I can quickly get the
translations for a given tweet without needing to send multiple request
to the database.